Made website fully responsive, updated fonts, and added UX fixes.
Added title and description for the product, fixed the mobile part of the website, added text that moves with the size of website, same for the buttons and images. Carousel changes size on website size too. Basically made the entire web page responsive towards mobile and ipad, tested both on a localhost devices. Changed the font on all the text places on the website, made all buttons selectable and added a quantity selector of maximum 10

-tliempt